refactor: change userModel.ts file name
