+++ src/4-framework/ioc/modelsModule.ts
@@ -1,6 +1,6 @@
 import { ContainerModule, interfaces } from 'inversify'
 
-import { UserModel } from '../models/userModal'
+import { UserModel } from '../models/userModel'
 
 export const ModelsModule = new ContainerModule((bind: interfaces.Bind) => {
   bind<UserModel>(UserModel).toConstructor(UserModel)
